Impersonate users via api as well
